+++ internal/lang/eval/internal/evalglue/glue.go
@@ -38,5 +38,5 @@ type Glue interface {
 	// resource instance object, but guaranteed to have already been validated.
 	// The implementation of this method should not call ConfigValue again
 	// and should instead just trust the value given as an argument.
-	ResourceInstanceValue(ctx context.Context, ri *configgraph.ResourceInstance, configVal cty.Value, providerInst configgraph.Maybe[*configgraph.ProviderInstance]) (cty.Value, tfdiags.Diagnostics)
+	ResourceInstanceValue(ctx context.Context, ri *configgraph.ResourceInstance, configVal cty.Value, providerInst configgraph.Maybe[*configgraph.ProviderInstance], riDeps addrs.Set[addrs.AbsResourceInstance]) (cty.Value, tfdiags.Diagnostics)
 }